refactor[#589]: 회원가입 폼에 이전 페이지로 이동하는 모달 로직 추가
